Fix overflow bug on 32-bit machines
Commit UUID: 292b1644-fb99-4f88-8d95-7cbb16597222

+++ trunk/ebml.c
@@ -53,7 +53,7 @@ struct elem_stack_ent {
     const struct elem_data  *data;
     size_t                  hdrlen;
     uint64_t                elen;
-    size_t                  totlen;
+    uint64_t                totlen;
     unsigned                segment;
     struct buf              *buf;
     ebml_master_cb_t        *master_cb;
